Add optional support for rDNS resolution via the c-ares library.
Passing '--with-c-ares' to configure enables it, in which case
thread_resolve.[ch] are skipped in favor of the library's resolver
routines.

Note that using this means that reverse name lookup will be strictly
DNS-based and thus won't take things like /etc/hosts or NIS into
account (hence it being disabled by default).

[See: http://c-ares.haxx.se/]