More structured preferences.
Less repetitive boilerplate in the preference loading/saving code,
fewer things to remember to update when adding a new one.  Just add a
member to pref_struct and a corresponding entry in the preferences
array and the infrastructure will take care of the rest.

Also enforces a distinction between *preferences*, which are saved
persistently in the config file, and command-line settings, which are
not.  Thus the 'stationary' and 'position' settings are removed from
pref_struct.